Editor/Extractor/requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.7.4
fastapi==0.110.0
uvicorn==0.27.1
numpy==1.24.4
